requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
supabase-py==0.6.0
gunicorn==22.0.0